fix: bound the rerun selection for both schedulers, not one
The cap on an explicit test case list lived in the open source enqueuer.
The control plane runs its own scheduler and had none, so the limit applied
to exactly half the deployments that could hit it - which is to say it did
not exist.

Both schedulers do call the shared ValidateExecutionRequest, so the check
moves there and the enqueuer's copy goes. Its test moves with it, plus one
asserting the request entry point applies it rather than only the helper -
that being the whole point of the move.

Refusing rather than truncating is unchanged and worth restating: a caller
who asked for five thousand specific tests and silently got a thousand
would read the result as those tests having passed.
